[vim] (journal) Fix bug: task is not initiating...
It was an occasional bug. Sometimes the mapping worked nicely, but sometimes it just didn't.

The problem was when there was no empty line at the end of the file. When opening 2 empty lines and moving the cursor around
the paragraphs, there was no room for deleting until the folding end.

The solution: open more lines, e.g, 4.